+++ src/pages/confide/[id].tsx
@@ -1,79 +1,73 @@
-import { useState, useEffect, useRef } from 'react';
-import { useRouter } from 'next/router';
+import { useMemo } from 'react';
 import { getPageInfo } from '@/services/api/getPageInfo';
 import { getContentPage } from '@/services/api/getContentPage';
-import { GetPageInfoResponseInterface } from '@/services/api/getPageInfo';
 import { GetContentPageResponseInterface } from '@/services/api/getContentPage';
+import { GetPageInfoResponseInterface } from '@/services/api/getPageInfo';
 import MainLayout from '@/components/Layout/MainLayout';
-import ConfideContentSkeleton from '@/components/ConfideContentSkeleton';
 import dayjs from 'dayjs';
 import { renderContentConfidePage } from '@/components/pages/ContentPage';
 import React from 'react';
 import AlertError from '@/components/Alert/AlertError';
+import { jsonDecode } from '@/helpers/urlHelper';
 
-export async function getServerSideProps() {
-  return {
-    props: {},
-  };
+export async function getServerSideProps(context: any) {
+  const pageId = context?.params?.id ?? '';
+  try{
+    const [pageInfoResp, pageContentResp] = await Promise.all([
+      getPageInfo({ page_id: pageId }), getContentPage({ block_id: pageId })
+    ])
+
+    return {
+      props: {
+        respErr: null,
+        pageInfoResp,
+        pageContentResp
+      },
+    };
+  }catch(e: any) {
+    const respErr = jsonDecode(e.body);
+    return {
+      props: {
+        respErr: respErr,
+        pageInfoResp: null,
+        pageContentResp: null
+      },
+    };
+  }
 }
 
 function ConfidePage(props: any) {
-  const router = useRouter();
-  const { id } = router.query;
-  const pageId = id as string;
+  const { pageInfoResp, pageContentResp, respErr } = props;
+  const pageInfo: GetPageInfoResponseInterface = useMemo(() => {
+    return pageInfoResp;
+  }, [pageInfoResp]);
 
-  const [paragraph, setParagraph] = useState<GetContentPageResponseInterface>();
-  const [pageInfo, setPageInfo] = useState<GetPageInfoResponseInterface>();
-  const [isError, setIsError] = useState<boolean>(false);
-  const [isLoading, setIsLoading] = useState<boolean>(false);
+  const contentPageInfo: GetContentPageResponseInterface = useMemo(() => {
+    return pageContentResp;
+  }, [pageContentResp]);
 
-  const isComponentMounted = useRef(false);
-  useEffect(() => {
-    isComponentMounted.current = true;
+  const pageTitle = useMemo(() => {
+    return pageInfo?.properties?.title?.rich_text[0]?.plain_text ?? '';
+  }, [pageInfo]);
 
-    return () => {
-      isComponentMounted.current = false;
-    };
-  }, []);
+  const tagsList = useMemo(() => {
+    return pageInfo?.properties?.tags?.multi_select?.map((item, idx) => {
+      return (
+        <span
+          key={idx + 'idx'}
+          style={{ zIndex: 1, color: item.color, marginRight: 4 }}
+        >
+          {item.name}
+        </span>
+      );
+    })
+  }, [pageInfo]);
 
-  useEffect(() => {
-    if (!pageId || !isComponentMounted.current) return;
-    setIsLoading(true);
+  const createDate = useMemo(() => {
+    return dayjs(pageInfo?.created_time ?? 0).unix();
+  }, [pageInfo])
 
-    getPageInfo({ page_id: pageId })
-      .then((res) => {
-        if (!isComponentMounted.current) return;
-        setIsLoading(false);
-        setIsError(false);
-        setPageInfo(res);
-      })
-      .catch((err) => {
-        if (!isComponentMounted.current) return;
-        setIsLoading(false);
-        setIsError(true);
-        console.log('get page info err ...', err);
-      });
-
-    getContentPage({ block_id: pageId })
-      .then((res) => {
-        if (!isComponentMounted.current) return;
-        setIsLoading(false);
-        setIsError(false);
-        setParagraph(res);
-      })
-      .catch((err) => {
-        if (!isComponentMounted.current) return;
-        setIsLoading(false);
-        setIsError(true);
-        console.log('get content page err...', err);
-      });
-  }, [pageId]);
-
-  if (isLoading) {
-    return <ConfideContentSkeleton />;
-  }
-
-  if (isError) {
+  if (respErr !== null) {
     return (
       <div className='pt-8'>
         <AlertError />
@@ -81,25 +75,14 @@ function ConfidePage(props: any) {
     )
   }
 
-  const createDate = dayjs(pageInfo?.created_time ?? 0).unix();
-
   return (
-    <div className="pt-8 ">
+    <div className="pt-8">
       <div className="mb-4">
         <h2 className="text-3xl mb-4 text-slate-700 font-semibold">
-          {pageInfo?.properties?.title?.rich_text[0]?.plain_text ?? ''}
+          {pageTitle}
         </h2>
         <div className="flex">
-          {pageInfo?.properties?.tags?.multi_select?.map((item, idx) => {
-            return (
-              <span
-                key={idx + 'idx'}
-                style={{ zIndex: 1, color: item.color, marginRight: 4 }}
-              >
-                {item.name}
-              </span>
-            );
-          })}
+          {tagsList}
         </div>
         <div>
           <span className="text-sm gray-700">
@@ -107,8 +90,8 @@ function ConfidePage(props: any) {
           </span>
         </div>
       </div>
-      <div className="py-4">
-        {paragraph?.results?.map((item, idx) => {
+      <div className="py-2">
+        {contentPageInfo?.results?.map((item, idx) => {
           return (
             <React.Fragment key={idx}>
               {renderContentConfidePage(item.type, item)}