+++ src/pages/index.tsx
@@ -2,6 +2,7 @@ import HomeLayout from '@/components/Layout/HomeLayout';
 import BlogCard from '@/components/cards/BlogCard';
 import { getArticleList } from '@/services/api/getArticleList';
 import { useRouter } from 'next/router';
+import { generateConfideDetailRouter } from '@/services/bussiness/router/generateRouter';
 
 export async function getServerSideProps() {
   const reqParam = {
@@ -28,7 +29,10 @@ function HomePage(props: any) {
   const router = useRouter();
 
   const handleClickViewDetailPageItem = (id: string) => {
-    return router.push(`/confide/${id}`);
+    const asPath = generateConfideDetailRouter({ id: id })?.asPath;
+    if (asPath) {
+      router.push(asPath);
+    }
   };
 
   return (